requirements.txt
flask
tensorflow
pillow
numpy
Pillow
